Reorganizing state views/tests for recipient app
